Update worker environment via bashrc edits
The  attempts to control the work start up and use of TMPDIR via
the job script won't work because workers are started outside the
job context via ssh.

Include sample bashrc modificaitons to set up the TMPDIR for java
and provide other necessary configurations for spark workers, e.g.
the same python environment as the caller.